backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
requests
google-generativeai
